+++ RIProxy/IRouter.cs
@@ -7,6 +7,6 @@ namespace Gamma.Proxy
 {
     interface IRouter
     {
-        void Send(IPEndPoint source, short port, byte[] payload);
+        void Send(IPEndPoint source, int port, byte[] payload);
     }
 }
