+++ RIProxy/Router.cs
@@ -8,32 +8,32 @@ namespace Gamma.Proxy
 {
     class Router : IRouter
     {
-        protected ConcurrentDictionary<(IPEndPoint, short), (IPEndPoint Dest, short To)> Routes { get; }
+        protected ConcurrentDictionary<(IPEndPoint, int), (IPEndPoint Dest, int To)> Routes { get; }
         protected IPacketHandler Handler { get; }
         protected ISocketsHolder SocketsHolder { get; }
 
         public Router(ISocketsHolder socketsHolder, IPacketHandler packetHandler)
         {
-            Routes = new ConcurrentDictionary<(IPEndPoint, short), (IPEndPoint, short)>();
+            Routes = new ConcurrentDictionary<(IPEndPoint, int), (IPEndPoint, int)>();
             Handler = packetHandler;
             SocketsHolder = socketsHolder;
         }
 
-        public void Send(IPEndPoint source, short from, byte[] payload)
+        public void Send(IPEndPoint source, int from, byte[] payload)
         {
             if(Handler.Handle(source, from, payload))
             {
-                if (Routes.TryGetValue((source, from), out (IPEndPoint Dest, short To) route))
+                if (Routes.TryGetValue((source, from), out (IPEndPoint Dest, int To) route))
                 {
                     SocketsHolder.GetSocket(route.To).Send(route.Dest, payload);
                 }
             }
         }
 
-        public bool AddRoute(IPEndPoint source, short from, IPEndPoint dest, short to) =>
+        public bool AddRoute(IPEndPoint source, int from, IPEndPoint dest, int to) =>
             Routes.TryAdd((source, from), (dest, to));
 
-        public bool RemoveRoute(IPEndPoint source, short port, out (IPEndPoint, short) target) =>
+        public bool RemoveRoute(IPEndPoint source, int port, out (IPEndPoint, int) target) =>
             Routes.TryRemove((source, port), out target);
 
         public void ClearRoute() =>
